Add URLRequest extension to simplify usage
